Add files via upload
This project was my independent project which I undertook solely on the basis of my interest in finance and mathematics and I thoroughly enjoyed throughout the time I was involved in it. The project was about how big hedge funds and other fund management houses invest huge amount of money in the market. We took the mathematical models as a reference and then performed different operations using two variables mainly in consideration which was expected return and risk or we can also say mean and standard deviation of the returns. We also learnt about how markets move under the influence of different news pieces. Learnt about the finances as well different terms like shorting, margin, call, put, options of a stock and many more in depth of financial markets.